TIScam: Prevent VC++ x64 warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10951 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/TISCam/DynDialogs.cpp
@@ -140,7 +140,7 @@ int InputBox(HWND hwnd, LPCTSTR prompt, LPCTSTR title, LPTSTR buffer, INT buflen
     ret = DialogBoxIndirectParamA(NULL, (LPDLGTEMPLATE)hgbl, hwnd, 
                                  (DLGPROC)InputBoxDlgProc, (int)data); 
     GlobalFree(hgbl); 
-    return (ret > 0) ? ret : 0; 
+    return static_cast<int>((ret > 0) ? ret : 0); 
 } 
 //---------------------------------------------------------------------------
